Added some kind of picking
Picking might be working.  Currently it is set to pick the tile in the
middle of the screen.  Tin, you could add this command to your input
routines for handling touch events:

mRenderer.selectTile( pick ( screenX, screenY );

where screenX and screenY is where the player clicked, and maybe this
would cause the tile to be selected.